+++ src/store/src/Bridge/ChromaDb/Store.php
@@ -18,6 +18,7 @@ use Codewithkyrian\ChromaDB\Responses\QueryItemsResponse;
 use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\Metadata;
 use Symfony\AI\Store\Document\VectorDocument;
+use Symfony\AI\Store\Document\VectorDocumentInterface;
 use Symfony\AI\Store\Exception\InvalidArgumentException;
 use Symfony\AI\Store\Exception\RuntimeException;
 use Symfony\AI\Store\Exception\UnsupportedQueryTypeException;
@@ -59,9 +60,9 @@ final class Store implements ManagedStoreInterface, StoreInterface
         }
     }
 
-    public function add(VectorDocument|array $documents): void
+    public function add(VectorDocumentInterface|array $documents): void
     {
-        if ($documents instanceof VectorDocument) {
+        if ($documents instanceof VectorDocumentInterface) {
             $documents = [$documents];
         }
 
@@ -171,7 +172,7 @@ final class Store implements ManagedStoreInterface, StoreInterface
     /**
      * @param array{where?: array<string, string>, whereDocument?: array<string, mixed>, include?: array<string>, limit?: positive-int} $options
      *
-     * @return iterable<VectorDocument>
+     * @return iterable<VectorDocumentInterface>
      */
     private function queryVector(VectorQuery $query, array $options): iterable
     {
@@ -192,7 +193,7 @@ final class Store implements ManagedStoreInterface, StoreInterface
     /**
      * @param array{where?: array<string, string>, whereDocument?: array<string, mixed>, include?: array<string>, limit?: positive-int} $options
      *
-     * @return iterable<VectorDocument>
+     * @return iterable<VectorDocumentInterface>
      */
     private function queryText(TextQuery $query, array $options): iterable
     {
@@ -226,7 +227,7 @@ final class Store implements ManagedStoreInterface, StoreInterface
     }
 
     /**
-     * @return iterable<VectorDocument>
+     * @return iterable<VectorDocumentInterface>
      */
     private function transformResponse(QueryItemsResponse $queryResponse): iterable
     {